Allow to define gem in the gemfile with the path option
